entrando no processo de finalizacao

+++ app/Http/Controllers/Api/admin/modulos/validacaoFormulas/ExercicioVFController.php
@@ -9,6 +9,7 @@ use App\Http\Controllers\LogicLive\config\Configuracao;
 use App\Http\Controllers\LogicLive\modulos\validacaoFormulas\ExercicioVF;
 use App\NivelMVFLP;
 use App\Recompensa;
+use App\Resposta;
 use Illuminate\Http\Request;
 
 
@@ -44,27 +45,12 @@ class ExercicioVFController extends Controller
         try{
             // Verifica sê existe uma recompensa com o id da requisicao
             $recompensas=Recompensa::where('id', $request->id_recompensa)->get(); 
-            if(count($recompensas)==0){return response()->json(['success' => false, 'msg'=>'Recompensa não cadastrada', 'data'=>''],500);}
+            if(count($recompensas)==0){return response()->json(['success' => false, 'msg'=>'recompensa informada nao foi encontrada', 'data'=>''],500);}
 
             // Verifica sê existe um nivel com o id da requisicao
             $nivel=NivelMVFLP::where('id',$request->id_nivel['id'])->get(); 
-            if(count($nivel)==0){return response()->json(['success' => false, 'msg'=>'Nivel não cadastrado', 'data'=>''],500);}
+            if(count($nivel)==0){return response()->json(['success' => false, 'msg'=>'nivel informado nao foi encontrado', 'data'=>''],500);}
 
-
-
-            $exercicio->id_recompensa=$request->id_recompensa['id'];
-            $exercicio->id_nivel=$request->id_nivel['id'];
-            $exercicio->nome=$request->nome;
-            $exercicio->enunciado=$request->enunciado;
-        
-            $exercicio->tempo=$request->tempo;
-            
-            $exercicio->descricao=$request->descricao;
-            $exercicio->ativo=$request->ativo;
-
-            $exercicio->qndt_erros=$request->qndt_erros;
-            $exercicio->hash='';
-            $exercicio->url='';
 
             $formula = new Formula();
             $formula->formula =$request->id_formula["formula"];
@@ -75,41 +61,53 @@ class ExercicioVFController extends Controller
             $formula->iniciar_zerada =$request->id_formula["iniciar_zerada"];
             $formula->inicio_personalizado =$request->id_formula["inicio_personalizado"];
             $formula->inicializacao_completa =$request->id_formula["inicializacao_completa"];
+            $formula->save();
+            $exercicio->id_recompensa=$request->id_recompensa['id'];
+            $exercicio->id_nivel=$request->id_nivel['id'];
+            $exercicio->nome=$request->nome;
+            $exercicio->enunciado=$request->enunciado;
+            $exercicio->tempo=$request->tempo;      
+            $exercicio->descricao=$request->descricao;
+            $exercicio->ativo=$request->ativo;
+            $exercicio->qndt_erros=$request->qndt_erros;
+            $exercicio->hash='';
+            $exercicio->url='';
+            $exercicio->id_formula=$formula->id;
+            $exercicio->save();
+            $exercicio->url=$this->config->urlExercicioValidacao().$exercicio->id;
+            
             if($request->id_formula["inicio_personalizado"]==true && $request->id_formula["iniciar_zerada"]==false ){
                 $formula->lista_passos =json_encode ($request->id_formula["lista_passos"]);
                 $formula->lista_derivacoes =json_encode ($request->id_formula["lista_derivacoes"]);
                 $formula->lista_ticagem =json_encode ($request->id_formula["lista_ticagem"]);
                 $formula->lista_fechamento =json_encode ($request->id_formula["lista_fechamento"]);
+                $formula->save();
             }
 
-            $formula->save();
-            $exercicio->id_formula=$formula->id;
-            $exercicio->save();
-            
-
-
             if($this->config->ativo()){        
-                $criadoLogicLive = $this->logicLive_exercicio->criarNivel([
+                $criadoLogicLive = $this->logicLive_exercicio->criarExercicio([
                     'rec_codigo'=>$recompensas[0]->id_logic_live, 
                     'niv_codigo'=>$nivel[0]->meu_id_logic_live, 
                     'exe_tempoexecucao'=> $request->tempo,
-                    'exe_link'=>$this->config->linkHospedagem().$exercicio->id,
+                    'exe_link'=>$exercicio->url,
                     'exe_nome'=>$request->nome,
                     'exe_descricao'=>$request->descricao,
                     'exe_ativo'=>$request->ativo
                 ]);
-
-                if($criadoLogicLive['success']=false){
+                if($criadoLogicLive['success']==false){
                     return response()->json(['success' => false, 'msg'=>$criadoLogicLive['msg'], 'data'=>''],500);
                     $exercicio->delete();
+                    $formula->delete();
                 }
+                $exercicio->hash= $criadoLogicLive['data']['exe_hash'];
+                $exercicio->id_logic_live= $criadoLogicLive['data']['exe_codigo'];
             }
-            $exercicio->hash= $criadoLogicLive['data']['exe_hash'];
+
             $exercicio->save();
 
-            return response()->json(['success' => true, 'msg'=>'Cadastrado!', 'data'=>'']);
+            return response()->json(['success' => true, 'msg'=>'exercicio criado no banco de dados', 'data'=>'']);
         }catch(\Exception $e){
-            return response()->json(['success' => false, 'msg'=>$e, 'data'=>''],500);
+            return response()->json(['success' => false, 'msg'=>'erro no servidor', 'data'=>''],500);
         }
       
     }
@@ -122,7 +120,10 @@ class ExercicioVFController extends Controller
      */
     public function show($id)
     {
-        //
+        $exercicio = $this->exercicio->find($id);
+        $exercicio->id_formula = Formula::findOrFail($exercicio->id_formula);
+        $exercicio->id_recompensa = Recompensa::findOrFail($exercicio->id_recompensa);
+        return response()->json(['success' => true, 'msg'=>'', 'data'=> $exercicio]);
     }
 
     /**
@@ -134,7 +135,39 @@ class ExercicioVFController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
+        // try{
+
+            $exercicio = ExercicioMVFLP::findOrFail($id);
+            $exercicio->update($request->all());
+            $exercicio->save();
+           
+
+            if($this->config->ativo()){
+                $criadoLogicLive = $this->logicLive_exercicio->atualizarExercicio($exercicio->id_logic_live,[
+                    'rec_codigo'=>Recompensa::findOrFail($exercicio->id_recompensa)->id_logic_live, 
+                    'niv_codigo'=>NivelMVFLP::findOrFail($exercicio->id_nivel)->meu_id_logic_live, 
+                    'exe_tempoexecucao'=> 111,
+                    'exe_link'=>$exercicio->url,
+                    'exe_nome'=>$exercicio->nome,
+                    'exe_descricao'=>$exercicio->descricao,
+                    'exe_ativo'=>$exercicio->ativo
+                    ]);
+                // return  response()->json($exercicio->meu_id_logic_live);
+                if($criadoLogicLive['success']==false){
+                   
+                    return response()->json(['success' => false, 'msg'=>$criadoLogicLive['msg'], 'data'=>''],500);
+                }
+
+            }
+
+            return response()->json(['success' => true, 'msg'=>'exercicio atualizado no banco de dados', 'data'=>''], 200);
+     
+            
+        
+        // }catch(\Exception $e){
+        //     return response()->json(['success' => false, 'msg'=>'erro no servidor', 'data'=>''],500);
+        // }
+       
     }
 
     /**
@@ -147,14 +180,28 @@ class ExercicioVFController extends Controller
     {
         try{
             $exercicio = ExercicioMVFLP::findOrFail($id);
-
             $formula= Formula::findOrFail($exercicio->id_formula); 
-            $exercicio->delete();
+
+            // Caso exista uma resposta para o exercicio não deixa ele ser excluido
+            if(count(Resposta::where('id_exercicio','=', $exercicio->id)->get())!=0){
+                return response()->json(['success' => false, 'msg'=>'existe resposta para esse exercicio', 'data'=>''],500);
+            }
+            
+            if($this->config->ativo()){
+                $criadoLogicLive = $this->logicLive_exercicio->deletarExercicio($exercicio->id_logic_live);
+                if($criadoLogicLive['success']==false){
+                   
+                    return response()->json(['success' => false, 'msg'=>$criadoLogicLive['msg'], 'data'=>''],500);
+                }
+            }
+
             $formula->delete();
-            return response()->json(['success' => true, 'msg'=>'Niviel ('.$exercicio->nome.') deletado com sucesso', 'data'=>''], 200);
+            $exercicio->delete();
+            
+            return response()->json(['success' => true, 'msg'=>'exercicio deletado do banco de dados', 'data'=>''], 200);
         
         }catch(\Exception $e){
-            return response()->json(['success' => false, 'msg'=>$e, 'data'=>''],500);
+            return response()->json(['success' => false, 'msg'=>'erro no servidor', 'data'=>''],500);
         }
     
     }
